made ws server write operations truly asynchronous separated WebSocket interface and implementation introduced WebSocketFactory to allow users to better customize their websockets
